lab11_joins/pyproject.toml
[tool.poetry]
name = "lab11-joins"
version = "0.1.0"
description = ""
packages = [{include = "src"}]
authors = ["Anatoly Zelenin"]

[tool.poetry.dependencies]
python = "^3.10"
faust-streaming = {extras = ["rocksdb"], version = "^0.9.5"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
